# Speedup ```Benchmark.run```