--- a/inductiva/benchmarks/benchmark.py
+++ b/inductiva/benchmarks/benchmark.py
@@ -148,15 +148,16 @@ class Benchmark(Project):
             Self: The current instance for method chaining.
         """
         def _run(params):
-            simulator, input_dir, machine_group, kwargs = params
-            if not machine_group.started:
-                machine_group.start(wait_for_quotas=wait_for_quotas)
-            for _ in range(num_repeats):
-                simulator.run(input_dir=input_dir,
-                              on=machine_group,
-                              **kwargs)
+            with self:
+                simulator, input_dir, machine_group, kwargs = params
+                if not machine_group.started:
+                    machine_group.start(wait_for_quotas=wait_for_quotas)
+                for _ in range(num_repeats):
+                    simulator.run(input_dir=input_dir,
+                                  on=machine_group,
+                                  **kwargs)
 
-        with self, ThreadPoolExecutor() as executor:
+        with ThreadPoolExecutor() as executor:
             _ = executor.map(_run, self.runs)
         self.runs.clear()
         return self