--- a/inductiva/benchmarks/benchmark.py
+++ b/inductiva/benchmarks/benchmark.py
@@ -3,16 +3,12 @@ import enum
 import json
 import csv
 import logging
+import concurrent.futures
 from typing import Optional, Union
 from typing_extensions import Self
-from inductiva import types, resources
-from inductiva.simulators import Simulator
-from inductiva.projects import Project
-from inductiva.client import ApiException
-from inductiva.resources.machine_types import ProviderType
-from inductiva.utils.format_utils import CURRENCY_SYMBOL, TIME_UNIT
 from collections import defaultdict
-from concurrent.futures import ThreadPoolExecutor
+from inductiva import types, resources, projects, simulators, client
+from inductiva.utils.format_utils import CURRENCY_SYMBOL, TIME_UNIT
 
 
 class ExportFormat(enum.Enum):
@@ -30,7 +26,7 @@ class SelectMode(enum.Enum):
     DISTINCT = "distinct"
 
 
-class Benchmark(Project):
+class Benchmark(projects.Project):
     """Represents the benchmark runner."""
 
     class InfoKey:
@@ -59,7 +55,7 @@ class Benchmark(Project):
 
     def set_default(
         self,
-        simulator: Optional[Simulator] = None,
+        simulator: Optional[simulators.Simulator] = None,
         input_dir: Optional[str] = None,
         on: Optional[types.ComputationalResources] = None,
         **kwargs,
@@ -94,7 +90,7 @@ class Benchmark(Project):
 
     def add_run(
         self,
-        simulator: Optional[Simulator] = None,
+        simulator: Optional[simulators.Simulator] = None,
         input_dir: Optional[str] = None,
         on: Optional[types.ComputationalResources] = None,
         **kwargs,
@@ -158,7 +154,7 @@ class Benchmark(Project):
                                   on=machine_group,
                                   **kwargs)
 
-        with ThreadPoolExecutor() as executor:
+        with concurrent.futures.ThreadPoolExecutor() as executor:
             _ = executor.map(_run, self.runs)
         self.runs.clear()
         return self
@@ -281,7 +277,7 @@ class Benchmark(Project):
         """
 
         def _handle_suffix(executer):
-            if executer.host_type == ProviderType.GCP:
+            if executer.host_type == resources.machine_types.ProviderType.GCP:
                 return "-".join(executer.vm_name.split("-")[:-1])
             return executer.vm_name
 
@@ -296,7 +292,7 @@ class Benchmark(Project):
                 continue
             try:
                 machine.terminate(verbose=False)
-            except ApiException as api_exception:
+            except client.ApiException as api_exception:
                 logging.warning(api_exception)
 
         return self