--- a/inductiva/resources/machines_base.py
+++ b/inductiva/resources/machines_base.py
@@ -365,9 +365,10 @@ class BaseMachineGroup(ABC):
         self._api.start_vm_group(body=request_body)
         creation_time = format_utils.seconds_formatter(time.time() - start_time)
         self._started = True
+        quota_usage_table_str = self.quota_usage_table_str("used by resource")
         logging.info(f"{self} successfully started in {creation_time}.\n\n"
                      "The machine group is using the following quotas:\n"
-                     f"{self.quota_usage_table_str("used by resource")}")
+                     f"{quota_usage_table_str}")
         return True
 
     def terminate(self, verbose: bool = True, **kwargs):