This PR optimizes the performance of the ```Benchmark.run``` method by using multithreading to execute multiple API requests in parallel for starting machine groups and submitting tasks. Additionally, several logging statements that were previously spread across multiple locations have been consolidated into a single one. This change helps reduce log clutter caused by simultaneous execution of threads.

- Script to measure speedup:

```python
from inductiva.benchmarks import Benchmark
from inductiva.utils import download_from_url
from inductiva.simulators import AmrWind
from inductiva.resources import MachineGroup
import time

# Configure the benchmark with default parameters
input_dir = download_from_url(
    "https://storage.googleapis.com/inductiva-api-demo-files/"
    "amr-wind-input-example.zip",
    unzip=True)

benchmark = Benchmark(name="test-speedup-6") \
    .set_default(simulator=AmrWind(),
                 input_dir=input_dir,
                 sim_config_filename="abl_amd_wenoz.inp")

# Add multiple machines to measure execution time of the `run` method
machines = [
    MachineGroup("c2-standard-4"),
    MachineGroup("c2-standard-4"),
    MachineGroup("c2-standard-4"),
    MachineGroup("c2-standard-4"),
    MachineGroup("c2-standard-8"),
    MachineGroup("c2-standard-8"),
    MachineGroup("c2-standard-8"),
    MachineGroup("c2-standard-8"),
]

for machine in machines:
    benchmark.add_run(on=machine, n_vcpus=machine.n_vcpus.total)

# Measure the execution time of the `run` method
start = time.perf_counter()
benchmark.run(num_repeats=2)
end = time.perf_counter()

print("Exec. Time: {} seconds".format(round(end - start, 2)))
```

- Execution time results (on MacBook Air M3):

|          | **1 Machine (1st Run)** | **1 Machine (2nd Run)** | **4 Machines (1st Run)** | **4 Machines (2nd Run)** | **8 Machines (1st Run)** | **8 Machines (2nd Run)** |
|----------------------------|---------------------------|----------------------------|----------------------------|-----------------------------|----------------------------|-----------------------------|
| **Single Thread**           | 34.69s             | 34.06s              | 150.05s             | 154.08s              | 305.07s             | 322.25s              |
| **Multi Thread**            | 34.52s             | 33.93s              | 45.09s              | 45.61s               | 47.55s              | 48.62s               |
| **Speedup**                 | 1.01x                     | 1.00x                      | 3.33x                      | 3.38x                       | 6.42x                      | 6.63x                       |